+++ GeonBit.UI/Source/Entities/CheckBox.cs
@@ -96,7 +96,7 @@ namespace GeonBit.UI.Entities
         {
             EntityState state = _entityState;
             if (state != EntityState.MouseDown && Checked) { state = EntityState.MouseDown; }
-            return Resources.CheckBoxTextures[(int)state];
+            return Resources.CheckBoxTextures[state];
         }
 
         /// <summary>
